fix: rename components to improve consistency
- Rename components to improve consistency
  - Change `links-card` to `LinksCard`
  - Change `person` to `Person`
- Improve `LinksCard` component directory structure
  - Move images to `images` subdirectory
- Improve code documentation
