+++ pandemia-fi/src/components/spinner/SpinnerBlock.md
@@ -1,0 +1,5 @@
+SpinnerBlock example:
+
+```js
+<SpinnerBlock pizza="🍕" />
+```
