+++ pandemia-fi/src/pages/about/Person.md
@@ -1,0 +1,5 @@
+Person example:
+
+```js
+<Person name="Firstname Lastname">Example introduction</Person>
+```
