+++ pandemia-fi/src/pages/about/about.tsx
@@ -1,8 +1,12 @@
 import React from "react";
 import { Card, Flex, Box, Heading, Text, Link } from "rebass";
 
-import Person from "./person";
+import Person from "./Person";
 
+/**
+ * About section of the website
+ * "Tietoa meistä"
+ */
 const About: React.FunctionComponent = () => {
   return (
     <Flex flexWrap="wrap" py={2}>
