+++ pandemia-fi/src/pages/landing/components/LinksCard/LinksCard.md
@@ -1,0 +1,5 @@
+LinksCard example:
+
+```js
+<LinksCard />
+```
