+++ pandemia-fi/src/pages/landing/components/LinksCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LinksCard";
