--- pandemia-fi/src/pages/landing/components/links-card/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./links-card";
